requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Django==4.1.7
django-environ==0.9.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
platformdirs==3.0.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
